fix(onnx): cast image array to `float32` to prevent inference failures
NumPy's type promotion rules for mixed-type operations (`uint8` and
`float32`) differ across versions (e.g., 1.x vs. 2.x). Subtracting a
`float32` mean array from a `uint8` image array can unexpectedly produce
a `float64` result.

Passing a `float64` tensor to ONNX Runtime models expecting `float32`
inputs causes silent execution failures, resulting in zero detections on
certain hardware backends.

Explicitly casting the image array to `np.float32` prior to
normalization ensures strict type adherence and stable cross-platform
ONNX inference.

+++ onnx_models.py
@@ -116,10 +116,14 @@ class OnnxModel(ABC):
         """
         image = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
         image = cv2.resize(image, self.image_size)
-        image = (image - self.mean) / self.std
+        # Explicit float32 cast is required: numpy's type-promotion rules
+        # for uint8 − float32 vary across numpy versions (1.x vs 2.x).
+        # Without the cast the result may be float64, which ONNX Runtime
+        # rejects silently on some backends, producing 0 detections.
+        image = ((image.astype(np.float32) - self.mean) / self.std)
         image = np.transpose(image, [2, 0, 1])
         image = np.expand_dims(image, axis=0)
-        return image
+        return image  # shape (1, 3, H, W), dtype float32
 
     def _get_input_output(self):
         inputs = self.sess.get_inputs()
